+++ AngularClient/src/main/shared/shared.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
-import { MomentModule } from "angular2-moment";
+import { MomentModule } from "ngx-moment";
 
 @NgModule({
   exports: [